Add `screenshots` lane option to avoid clean build
While it makes sense to deletes DerivedData when generating the final screenshots, to be sure to avoid relying on potentially stale cache, this means that creating screenshots will always need to rebuild the whole app from scratch first.

When iterating by trial-and-error — typically when someone needs to tweak the screenshots or debug the lanes until they get to the result they want — it's a big waste of time to do a clean build on each iteration while we typically only changed a quick thing to tweak the setup and retry to see how it looks.

Also, when we will ultimately run this lane on CI every night or so to ensure the Screenshots UI Test targets still compile and work (just to test the lane, as opposed to actually re-generate all the screenshots for all the locales with the intention of updating them in App Store), re-using the DerivedData from a build of the app made from a previous CI step would also be useful.